+++ Ui/Listing/DataProvider/Banner.php
@@ -1,0 +1,75 @@
+<?php
+/**
+ * Copyright (c) 2026. Volodymyr Hryvinskyi. All rights reserved.
+ * Author: Volodymyr Hryvinskyi <volodymyr@hryvinskyi.com>
+ * GitHub: https://github.com/hryvinskyi
+ */
+
+declare(strict_types=1);
+
+namespace Hryvinskyi\BannerSliderAdminUi\Ui\Listing\DataProvider;
+
+use Hryvinskyi\BannerSliderAdminUi\Api\DataProvider\PrepareDataProcessorInterface;
+use Magento\Framework\Api\FilterBuilder;
+use Magento\Framework\Api\Search\ReportingInterface;
+use Magento\Framework\Api\Search\SearchCriteriaBuilder;
+use Magento\Framework\App\RequestInterface;
+use Magento\Framework\View\Element\UiComponent\DataProvider\DataProvider;
+
+/**
+ * Banner listing data provider
+ */
+class Banner extends DataProvider
+{
+    /**
+     * @param string $name
+     * @param string $primaryFieldName
+     * @param string $requestFieldName
+     * @param ReportingInterface $reporting
+     * @param SearchCriteriaBuilder $searchCriteriaBuilder
+     * @param RequestInterface $request
+     * @param FilterBuilder $filterBuilder
+     * @param PrepareDataProcessorInterface $prepareDataProcessor
+     * @param array $meta
+     * @param array $data
+     */
+    public function __construct(
+        string $name,
+        string $primaryFieldName,
+        string $requestFieldName,
+        ReportingInterface $reporting,
+        SearchCriteriaBuilder $searchCriteriaBuilder,
+        RequestInterface $request,
+        FilterBuilder $filterBuilder,
+        private readonly PrepareDataProcessorInterface $prepareDataProcessor,
+        array $meta = [],
+        array $data = []
+    ) {
+        parent::__construct(
+            $name,
+            $primaryFieldName,
+            $requestFieldName,
+            $reporting,
+            $searchCriteriaBuilder,
+            $request,
+            $filterBuilder,
+            $meta,
+            $data
+        );
+    }
+
+    /**
+     * @inheritDoc
+     */
+    #[\Override]
+    public function getData(): array
+    {
+        $data = parent::getData();
+
+        foreach ($data['items'] as &$item) {
+            $this->prepareDataProcessor->execute($item);
+        }
+
+        return $data;
+    }
+}